Updated SidePanel
