+++ frontend/src/components/organisms/SidePanel.tsx
@@ -1,89 +1,82 @@
-import React from "react";
-import { Box, Drawer } from "@mui/material";
-import { useState } from "react";
+import React, { useState, useEffect } from "react";
 import Button from "@/components/atoms/Button";
-import data from "@/components/charts/data/mongolia-province-data.json";
-import {
-  extractData,
-  groupByLivestockAndYear,
-  extractYearRange,
-  createMultipleDatasets,
-} from "@/utils/helpers";
-import { Mongolia } from "@/utils/interfaces";
-import DropDown from "@/components/atoms/DropDown";
 import SearchBar from "@/components/molecules/SearchBar";
 import BarChart from "@/components/charts/barchart";
+import { Box, Drawer } from "@mui/material";
 
 const SidePanel = () => {
-  const organizedData: Mongolia = data;
-
-  const extractedDataWithKeys = extractData(organizedData);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const goats = createMultipleDatasets(extractedDataWithKeys, "Goat");
-
-  const yearRange = Array.from(extractYearRange(extractedDataWithKeys)).sort(
-    (a, b) => a - b
-  );
-
   const livestockTypes = ["Cattle", "Horse", "Goat", "Camel", "Sheep"];
-
+  const yearRange = [
+    2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013,
+    2014,
+  ]; // Full year range
   const [filteredData, setFilteredData] = useState<any[]>([]);
-
-  const [selectedYear, setSelectedYear] = useState<number | null>(
+  const [selectedYear, setSelectedYear] = useState<number>(
     Math.max(...yearRange)
-  );
-
+  ); // Default to the latest year
   const [isPanelOpen, setIsPanelOpen] = useState(false);
   const [selectedAimag, setSelectedAimag] = useState<string | null>(null);
+
+  // Fetch data for a specific aimag and year
+  const loadProvince = async (aimag: string, year: number) => {
+    try {
+      const response = await fetch(
+        `http://localhost:8080/api/province/${aimag}`
+      );
+      if (!response.ok) {
+        throw new Error("Failed to fetch province data");
+      }
+      const data = await response.json();
+
+      // Transform data into the format required by BarChart
+
+      const province_land_area = data["province_land_area"];
+      const province_herders = data["province_herders"];
+
+      const transformedData = livestockTypes.map((type) => ({
+        x: type,
+        y: data[`province_number_of_${type.toLowerCase()}`]?.[year] || 0, // Default to 0 if no data for the year
+      }));
+
+      // Update filteredData state
+      setFilteredData([
+        {
+          aimag: data.province_name,
+          data: transformedData,
+        },
+      ]);
+    } catch (error) {
+      console.error("Error fetching province data:", error);
+      setFilteredData([]); // Reset data on error
+    }
+  };
 
   const handlePanelToggle = () => {
     setIsPanelOpen(!isPanelOpen);
     setSelectedAimag(null);
+    setFilteredData([]);
   };
 
-  const handleSearchResult = (aimag: string | null) => {
-    if (aimag) {
-      setSelectedAimag(aimag);
-      const yearToUse = selectedYear || Math.max(...yearRange);
-
-      const filtered = groupByLivestockAndYear(
-        yearToUse,
-        livestockTypes,
-        extractedDataWithKeys.filter(
-          (dataset) =>
-            dataset.Aimag === aimag && dataset.Year.includes(yearToUse)
-        )
-      );
-      setFilteredData(filtered);
-    } else {
-      setSelectedAimag(null);
-      setFilteredData([]);
+  const handleProvinceSearch = (aimag: string | null) => {
+    setSelectedAimag(aimag);
+    if (aimag && selectedYear) {
+      loadProvince(aimag, selectedYear); // Fetch data when both aimag and year are set
     }
   };
 
-  const selectedOption: string | null = null;
-
-  const handleYearChange = (year: string | null) => {
-    if (year) {
-      const yearAsNumber = Number(year);
-      setSelectedYear(yearAsNumber);
-      const aimagToUse = selectedAimag;
-
-      const filtered = groupByLivestockAndYear(
-        yearAsNumber,
-        livestockTypes,
-        extractedDataWithKeys.filter(
-          (dataset) =>
-            (!aimagToUse || dataset.Aimag === aimagToUse) &&
-            dataset.Year.includes(yearAsNumber)
-        )
-      );
-      setFilteredData(filtered);
-    } else {
-      setSelectedYear(null);
-      setFilteredData([]);
+  const handleYearSlider = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const year = Number(event.target.value);
+    setSelectedYear(year);
+    if (selectedAimag) {
+      loadProvince(selectedAimag, year); // Fetch data when both aimag and year are set
     }
   };
+
+  useEffect(() => {
+    if (selectedAimag && selectedYear) {
+      loadProvince(selectedAimag, selectedYear); // Fetch data initially if values are set
+    }
+  }, [selectedAimag, selectedYear]);
 
   return (
     <div>
@@ -111,15 +104,16 @@ const SidePanel = () => {
                   onClick={handlePanelToggle}
                   label="Close"
                   sx={{ position: "absolute", top: 8, right: 8 }}
-                ></Button>
+                />
               </div>
+
               <div style={{ marginTop: "60px" }}>
                 <div style={{ display: "flex", alignItems: "center" }}>
                   <div style={{ flex: 11, marginRight: "1rem" }}>
-                    <SearchBar onSearch={handleSearchResult} />
+                    <SearchBar onSearch={handleProvinceSearch} />
                   </div>
                 </div>
-                {/* <ScatterLinePlot datasets={goats} livestock={"Goats"} /> */}
+
                 <div
                   style={{
                     display: "flex",
@@ -128,14 +122,18 @@ const SidePanel = () => {
                   }}
                 >
                   <div style={{ flex: 11, marginRight: "1rem" }}>
-                    <DropDown
-                      options={yearRange.map((year) => year.toString())}
-                      value={selectedOption}
-                      onChange={handleYearChange}
-                      label="Select a year"
+                    <label>Select a Year: {selectedYear}</label>
+                    <input
+                      type="range"
+                      min={Math.min(...yearRange)}
+                      max={Math.max(...yearRange)}
+                      value={selectedYear}
+                      onChange={handleYearSlider}
+                      aria-label="Select a year"
                     />
                   </div>
                 </div>
+
                 {selectedAimag && filteredData.length > 0 && (
                   <div>
                     <BarChart
