+++ frontend/src/pages/sidepanelpage.tsx
@@ -1,14 +1,12 @@
 import React from "react";
 import SidePanel from "@/components/organisms/SidePanel";
-import SidePain from "@/components/organisms/SidePain";
 
 /** An About page */
 
 const SidePanelPage = () => {
   return (
     <div>
-      {/*<SidePanel></SidePanel>*/}
-      <SidePain></SidePain>
+      <SidePanel></SidePanel>
     </div>
   );
 };
